+++ Coffee-Bearer/src/api.h
@@ -1,0 +1,9 @@
+#ifndef API_H
+#define API_H
+
+#include <ESPAsyncWebServer.h>
+
+extern AsyncWebServer server;
+void configurar_rotas_api();
+
+#endif // API_H
