+++ Coffee-Bearer/src/hardware.cpp
@@ -1,0 +1,8 @@
+#include "hardware.h"
+
+void hardware_init() {
+	pinMode(BUZZER_PIN, OUTPUT);
+	pinMode(RELAY_PIN, OUTPUT);
+	digitalWrite(RELAY_PIN, LOW);
+	SPI.begin();
+}
