+++ Coffee-Bearer/src/hardware.h
@@ -1,0 +1,13 @@
+
+#ifndef HARDWARE_H
+#define HARDWARE_H
+
+#include <Arduino.h>
+#include <SPI.h>
+
+#define BUZZER_PIN 2
+#define RELAY_PIN 3
+
+void hardware_init();
+
+#endif // HARDWARE_H
