+++ Coffee-Bearer/src/log.h
@@ -1,0 +1,9 @@
+#ifndef LOG_H
+#define LOG_H
+
+#include <Arduino.h>
+#include "SPIFFS.h"
+
+void registrar_log(String evento);
+
+#endif // LOG_H
