+++ Coffee-Bearer/src/serialcmd.h
@@ -1,0 +1,6 @@
+#ifndef SERIALCMD_H
+#define SERIALCMD_H
+
+void processar_comando_serial();
+
+#endif // SERIALCMD_H
