+++ Coffee-Bearer/src/som.cpp
@@ -1,0 +1,9 @@
+#include "som.h"
+
+void som_autorizado() { tone(BUZZER_PIN, 1200, 80); delay(100); tone(BUZZER_PIN, 1500, 80); }
+void som_negado() { tone(BUZZER_PIN, 200, 400); }
+void som_inicializacao() { tone(BUZZER_PIN, 800, 60); delay(80); tone(BUZZER_PIN, 1000, 60); delay(80); tone(BUZZER_PIN, 1200, 60); }
+void som_cafe_pronto() { tone(BUZZER_PIN, 1300, 50); delay(60); tone(BUZZER_PIN, 1300, 50); delay(60); tone(BUZZER_PIN, 1300, 50); delay(60); tone(BUZZER_PIN, 1600, 100); }
+void som_dados_salvos() { tone(BUZZER_PIN, 1800, 50); delay(60); tone(BUZZER_PIN, 1800, 50); }
+void som_sem_cafe() { tone(BUZZER_PIN, 440, 150); delay(160); tone(BUZZER_PIN, 440, 150); delay(160); tone(BUZZER_PIN, 440, 150); }
+void som_reabastecido() { tone(BUZZER_PIN, 1500, 80); delay(100); tone(BUZZER_PIN, 1800, 80); delay(100); tone(BUZZER_PIN, 2200, 120); }
